# packages/backend-core/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/crm/merge-proposals": {
      "get": {
        "operationId": "CrmMergeProposalsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmMergeProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/merge-proposals/{id}": {
      "get": {
        "operationId": "CrmMergeProposalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmMergeProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/merge-proposals/{id}/apply": {
      "post": {
        "operationId": "CrmMergeProposalsController_apply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmMergeProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/merge-proposals/{id}/dismiss": {
      "post": {
        "operationId": "CrmMergeProposalsController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissMergeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmMergeProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/segments": {
      "get": {
        "operationId": "CrmSegmentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "post": {
        "operationId": "CrmSegmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSegmentBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/segments/{id}": {
      "get": {
        "operationId": "CrmSegmentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "patch": {
        "operationId": "CrmSegmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSegmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "delete": {
        "operationId": "CrmSegmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/crm/segments/{id}/contacts": {
      "get": {
        "operationId": "CrmSegmentsController_listContacts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CrmSegments"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/outreach/proposals/{id}": {
      "get": {
        "operationId": "OutreachProposalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "patch": {
        "operationId": "OutreachProposalsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/outreach/proposals/{id}/dismiss": {
      "post": {
        "operationId": "OutreachProposalsController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissProposalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OutreachProposals"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/curation/jobs": {
      "get": {
        "operationId": "CuratorJobsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "post": {
        "operationId": "CuratorJobsController_enqueue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueJobBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/curation/jobs/{id}": {
      "get": {
        "operationId": "CuratorJobsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/curation/jobs/claim": {
      "post": {
        "operationId": "CuratorJobsController_claim",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimJobBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/curation/jobs/{id}/ack": {
      "post": {
        "operationId": "CuratorJobsController_ack",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckJobBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
    "/v1/curation/jobs/{id}/fail": {
      "post": {
        "operationId": "CuratorJobsController_fail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailJobBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CuratorJobs"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Munin",
    "description": "HTTP surface area for the Munin platform. Authenticate with a bearer token (admin API key or delegated end-user token) or a session cookie.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "munin_session"
      }
    },
    "schemas": {
...
      "DismissMergeBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "CreateSegmentBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "filter": {
            "type": "object",
            "properties": {
              "tagsAny": {
                "maxItems": 32,
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                }
              },
              "tagsAll": {
                "maxItems": 32,
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                }
              },
              "companyId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 64
              },
              "searchQuery": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "contactedSince": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "x-nestjs_zod-self-required": true
          }
        },
        "required": [
          "name",
          "filter"
        ]
      },
      "UpdateSegmentBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "filter": {
            "type": "object",
            "properties": {
              "tagsAny": {
                "maxItems": 32,
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                }
              },
              "tagsAll": {
                "maxItems": 32,
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                }
              },
              "companyId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 64
              },
              "searchQuery": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "contactedSince": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": false
          }
        }
      },
      "UpdateProposalBody": {
        "type": "object",
        "properties": {
          "draftSubject": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "draftBody": {
            "type": "string",
            "minLength": 1,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        }
      },
      "DismissProposalBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "EnqueueJobBody": {
        "type": "object",
        "properties": {
          "jobUri": {
            "type": "string",
            "minLength": 1
          },
          "userPrompt": {
            "type": "string",
            "minLength": 1
          },
          "sourceEventType": {
            "type": "string"
          },
          "sourceEventPayload": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "dedupeKey": {
            "type": "string",
            "minLength": 1
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "delaySeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 86400
          }
        },
        "required": [
          "jobUri",
          "userPrompt"
        ]
      },
      "ClaimJobBody": {
        "type": "object",
        "properties": {
          "holder": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 25
          },
          "leaseSeconds": {
            "type": "integer",
            "minimum": 30,
            "maximum": 3600
          }
        },
        "required": [
          "holder"
        ]
      },
      "AckJobBody": {
        "type": "object",
        "properties": {
          "replyText": {
            "type": "string"
          },
          "toolCalls": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "totalTokens": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        }
      },
      "FailJobBody": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          },
          "retryable": {
            "type": "boolean"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "failedStep": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}